requirements.txt
base58
base64
json
ecdsa
eciespy
ttn
sgl>=0.9.14
python-dotenv
requests
paho-mqtt
ConfigArgParse
